requirements.txt
aioice==0.7.6
aiortc==1.3.0
av==8.1.0
imgaug==0.4.0
numpy==1.23.0
Pillow==8.4.0
streamlit==1.4.0.
streamlit-webrtc>=0.34.1
torch==1.10.0
torchaudio==0.10.0
torchvision==0.11.1
opencv-python
ffmpeg 
libsm==0.1.0
libfsext-python==20220529
pyglview==1.3.4
opencv-python-headless
opencv_contrib_python
pyOpenSSL==22.0.0
pysqlite3
